Refactor : 구조 변경
